raj/requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
arid==0.1
asciitree==0.3.3
async-generator==1.10
attrs==20.3.0
augment-nd==0.1.1
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cycler==0.10.0
Cython==0.29.21
daisy==0.2.1
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fasteners==0.16
funlib.learn.torch==0.1
funlib.segment==0.1
funlib.show.neuroglancer==0.1
future==0.18.2
google-apitools==0.5.31
google-auth==1.25.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
gunpowder==1.2.1
h5py==3.1.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.0.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lsd==0.1
mahotas==1.4.11
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mlpack==3.4.2
nbclient==0.5.1
nbconvert==5.6.0
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
neuroglancer==2.20
notebook==6.1.4
numcodecs==0.7.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkg-resources==0.0.0
prometheus-client==0.9.0
prompt-toolkit==2.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sockjs-tornado==1.0.7
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.6.0
torchvision==0.7.0
tornado==5.1.1
tqdm==4.57.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
waterz==0.9.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zarr==2.6.1
zipp==3.4.0
